+++ Kernel/System/Ticket.pm
@@ -1945,7 +1945,6 @@ sub TicketQueueSet {
                     RecipientID           => $UserID,
                     CustomerMessageParams => {
                         Queue => $Queue,
-                        Body  => $Param{Comment} || '',
                     },
                     TicketID => $Param{TicketID},
                     UserID   => $Param{UserID},
@@ -4555,12 +4554,9 @@ sub TicketOwnerSet {
             $Self->SendAgentNotification(
                 Type                  => 'OwnerUpdate',
                 RecipientID           => $Param{NewUserID},
-                CustomerMessageParams => {
-                    %Param,
-                    Body => $Param{Comment} || '',
-                },
-                TicketID => $Param{TicketID},
-                UserID   => $Param{UserID},
+                CustomerMessageParams => \%Param,
+                TicketID              => $Param{TicketID},
+                UserID                => $Param{UserID},
             );
         }
     }